# swagger/spec.json
{"openapi":"3.0.0","info":{"title":"Spotit API","version":"1.0.0"},"paths":{"/amusementPark":{"get":{"summary":"Retrieve a list of amusement parks","description":"Retrieve a list of amusement parks from the database.","responses":{"200":{"description":"A list of amusement parks.","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"integer","description":"The amusement park ID.","example":1},"name":{"type":"string","description":"The name of the amusement park.","example":"Wonderland"},"location":{"type":"string","description":"The location of the amusement park.","example":"123 Fun St, Adventure City"},"attractions":{"type":"array","description":"List of attractions in the amusement park.","items":{"type":"string","example":"Roller Coaster"}}}}}}}}}}},"/cinema":{"get":{"summary":"Retrieve a list of cinemas","description":"Retrieve a list of cinemas from the database.","responses":{"200":{"description":"A list of cinemas.","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"integer","description":"The cinema ID.","example":1},"name":{"type":"string","description":"The name of the cinema.","example":"Cinema Paradiso"},"location":{"type":"string","description":"The location of the cinema.","example":"456 Movie Blvd, Film City"},"screens":{"type":"integer","description":"The number of screens in the cinema.","example":10}}}}}}}}}},"/login":{"post":{"summary":"User login","description":"Authenticate user and return a JWT token.","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string","description":"The user's email.","example":"alice.smith@gmail.com"},"password":{"type":"string","description":"The user's password.","example":"password123"}}}}}},"responses":{"200":{"description":"Successful login.","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string","description":"JWT token.","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},"email":{"type":"string","description":"The user's email.","example":"alice.smith@gmail.com"},"status":{"type":"string","description":"The user's status.","example":"user"}}}}}},"401":{"description":"Invalid email or password.","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","description":"Error message.","example":"Invalid email or password"}}}}}}}}},"/leaderboard":{"get":{"summary":"Retrieve the leaderboard","description":"Retrieve the leaderboard with user rankings.","responses":{"200":{"description":"A list of users with their rankings.","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"userId":{"type":"integer","description":"The user ID.","example":1},"username":{"type":"string","description":"The username.","example":"john_doe"},"points":{"type":"integer","description":"The user's points.","example":1500}}}}}}}}}},"/manager/users":{"get":{"summary":"Retrieve a list of users","description":"Retrieve a list of users from the database.","responses":{"200":{"description":"A list of users.","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"integer","description":"The user ID.","example":1},"email":{"type":"string","description":"The user's email.","example":"alice.smith@gmail.com"},"status":{"type":"string","description":"The user's status.","example":"user"}}}}}}}}}},"/manager/user":{"get":{"summary":"Retrieve a specific user","description":"Retrieve a specific user from the database.","responses":{"200":{"description":"A user object.","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","description":"The user ID.","example":1},"email":{"type":"string","description":"The user's email.","example":"alice.smith@gmail.com"},"status":{"type":"string","description":"The user's status.","example":"user"}}}}}}}},"delete":{"summary":"Delete a user","description":"Delete a user from the database.","responses":{"200":{"description":"User deleted successfully.","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","description":"Success message.","example":"User deleted successfully"}}}}}}}},"patch":{"summary":"Update a user","description":"Update a user's information.","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string","description":"The user's email.","example":"alice.smith@gmail.com"},"password":{"type":"string","description":"The user's password.","example":"newpassword123"}}}}}},"responses":{"200":{"description":"User updated successfully.","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","description":"The user ID.","example":1},"email":{"type":"string","description":"The user's email.","example":"alice.smith@gmail.com"},"status":{"type":"string","description":"The user's status.","example":"user"}}}}}}}},"post":{"summary":"Create a new user","description":"Create a new user in the database.","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string","description":"The user's email.","example":"alice.smith@gmail.com"},"password":{"type":"string","description":"The user's password.","example":"password123"}}}}}},"responses":{"201":{"description":"User created successfully.","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","description":"The user ID.","example":1},"email":{"type":"string","description":"The user's email.","example":"alice.smith@gmail.com"},"status":{"type":"string","description":"The user's status.","example":"user"}}}}}}}}},"/manager/userWithoutPassword":{"patch":{"summary":"Update a user without changing the password","description":"Update a user's information without changing the password.","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string","description":"The user's email.","example":"alice.smith@gmail.com"}}}}}},"responses":{"200":{"description":"User updated successfully.","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","description":"The user ID.","example":1},"email":{"type":"string","description":"The user's email.","example":"alice.smith@gmail.com"},"status":{"type":"string","description":"The user's status.","example":"user"}}}}}}}}},"/manager/userslist":{"get":{"summary":"Retrieve a list of user names","description":"Retrieve a list of user names from the database.","responses":{"200":{"description":"A list of user names.","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"integer","description":"The user ID.","example":1},"name":{"type":"string","description":"The user's name.","example":"Alice Smith"}}}}}}}}}},"/manager/vouchers":{"get":{"summary":"Retrieve a list of vouchers","description":"Retrieve a list of vouchers from the database.","responses":{"200":{"description":"A list of vouchers.","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"integer","description":"The voucher ID.","example":1},"name":{"type":"string","description":"The name of the voucher.","example":"Discount Voucher"},"discount":{"type":"number","description":"The discount percentage.","example":10}}}}}}}}}},"/manager/voucher":{"get":{"summary":"Retrieve a specific voucher","description":"Retrieve a specific voucher from the database.","responses":{"200":{"description":"A voucher object.","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","description":"The voucher ID.","example":1},"name":{"type":"string","description":"The name of the voucher.","example":"Discount Voucher"},"discount":{"type":"number","description":"The discount percentage.","example":10}}}}}}}},"post":{"summary":"Create a new voucher","description":"Create a new voucher in the database.","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","description":"The name of the voucher.","example":"Discount Voucher"},"discount":{"type":"number","description":"The discount percentage.","example":10}}}}}},"responses":{"201":{"description":"Voucher created successfully.","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","description":"The voucher ID.","example":1},"name":{"type":"string","description":"The name of the voucher.","example":"Discount Voucher"},"discount":{"type":"number","description":"The discount percentage.","example":10}}}}}}}},"delete":{"summary":"Delete a voucher","description":"Delete a voucher from the database.","responses":{"200":{"description":"Voucher deleted successfully.","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","description":"Success message.","example":"Voucher deleted successfully"}}}}}}}},"patch":{"summary":"Update a voucher","description":"Update a voucher's information.","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","description":"The name of the voucher.","example":"Discount Voucher"},"discount":{"type":"number","description":"The discount percentage.","example":10}}}}}},"responses":{"200":{"description":"Voucher updated successfully.","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","description":"The voucher ID.","example":1},"name":{"type":"string","description":"The name of the voucher.","example":"Discount Voucher"},"discount":{"type":"number","description":"The discount percentage.","example":10}}}}}}}}},"/reports/{id}":{"get":{"summary":"Obtenir un rapport par ID","tags":["Reports"],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"},"description":"ID du rapport"}],"responses":{"200":{"description":"Rapport obtenu avec succès","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}}}}},"404":{"description":"Rapport non trouvé"}}}},"/reports":{"get":{"summary":"Récupérer les rapports dans une région donnée","tags":["Reports"],"parameters":[{"in":"query","name":"region","required":true,"schema":{"type":"string"},"description":"Région des rapports"}],"responses":{"200":{"description":"Liste des rapports obtenue","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}}}}}}}},"post":{"summary":"Ajouter un rapport avec une image","tags":["Reports"],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"image":{"type":"string","format":"binary"},"title":{"type":"string"},"description":{"type":"string"}}}}}},"responses":{"201":{"description":"Rapport ajouté avec succès","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string"},"message":{"type":"string"}}}}}},"400":{"description":"Données invalides"}}}},"/reportType":{"get":{"summary":"Retrieve a list of report types","description":"Retrieve a list of report types from the database.","responses":{"200":{"description":"A list of report types.","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"integer","description":"The report type ID.","example":1},"name":{"type":"string","description":"The name of the report type.","example":"Bug Report"}}}}}}}}}},"/restaurant":{"get":{"summary":"Retrieve a list of restaurants","description":"Retrieve a list of restaurants from the database.","responses":{"200":{"description":"A list of restaurants.","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"integer","description":"The restaurant ID.","example":1},"name":{"type":"string","description":"The name of the restaurant.","example":"Le Gourmet"},"address":{"type":"string","description":"The address of the restaurant.","example":"123 Main St, Cityville"},"cuisine":{"type":"string","description":"The type of cuisine the restaurant offers.","example":"French"}}}}}}}}}},"/user/me":{"get":{"summary":"Retrieve the current user's information","description":"Retrieve the current user's information based on the JWT token.","responses":{"200":{"description":"The user's information.","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","description":"The user ID.","example":1},"email":{"type":"string","description":"The user's email.","example":"alice.smith@gmail.com"},"status":{"type":"string","description":"The user's status.","example":"user"}}}}}}}},"patch":{"summary":"Update the current user's information","description":"Update the current user's information based on the JWT token.","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string","description":"The user's email.","example":"alice.smith@gmail.com"},"password":{"type":"string","description":"The user's password.","example":"newpassword123"}}}}}},"responses":{"200":{"description":"The user's updated information.","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","description":"The user ID.","example":1},"email":{"type":"string","description":"The user's email.","example":"alice.smith@gmail.com"},"status":{"type":"string","description":"The user's status.","example":"user"}}}}}}}}},"/user/meWithoutPassword":{"patch":{"summary":"Update the current user's information without changing the password","description":"Update the current user's information without changing the password based on the JWT token.","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string","description":"The user's email.","example":"alice.smith@gmail.com"}}}}}},"responses":{"200":{"description":"The user's updated information.","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","description":"The user ID.","example":1},"email":{"type":"string","description":"The user's email.","example":"alice.smith@gmail.com"},"status":{"type":"string","description":"The user's status.","example":"user"}}}}}}}}}},"components":{},"tags":[{"name":"Reports","description":"API pour gérer les rapports"}]}